RET-5698 (#614)
