+++ src/test/unit/controllers/ApplicationDetailsController.test.ts
@@ -33,7 +33,7 @@ describe('Application Details Controller', () => {
         TranslationKeys.APPLICATION_DETAILS,
         expect.objectContaining({
           applicationType: 'Amend my response',
-          isRespondButton: true,
+          isAdminRespondButton: true,
           respondRedirectUrl: '/respond-to-application/5d0118c9-bdd6-4d32-9131-6aa6f5ec718e?lng=en',
         })
       );
